+++ nuxt.config.ts
@@ -34,5 +34,12 @@ export default defineNuxtConfig({
     },
   },
 
+  runtimeConfig: {
+    public: {
+      supabaseUrl: process.env.SUPABASE_URL || '',
+      supabaseKey: process.env.SUPABASE_KEY || '',
+    },
+  },
+
   compatibilityDate: '2025-03-29',
 })